# Return consistent table type