--- a/src/linear_combination.jl
+++ b/src/linear_combination.jl
@@ -53,6 +53,8 @@ are specified, then the [`LinearCombination`](@ref) is applied to all columns.
 function apply(table, LC::LinearCombination; cols=nothing)
     Tables.istable(table) || throw(MethodError(apply, (table, LC)))
 
+    T = Tables.materializer(table)
+
     cols = _to_vec(cols)  # handle single column name
 
     # Error if dimensions don't match
@@ -61,10 +63,14 @@ function apply(table, LC::LinearCombination; cols=nothing)
 
     # Keep the generic form when not specifying column names
     # because that is much more performant than selecting each col by name
-    cols === nothing && return [_sum_row(row, LC.coefficients) for row in Tables.rows(table)]
-
-    return [
-        _sum_row([row[cname] for cname in cols], LC.coefficients)
-        for row in Tables.rows(table)
-    ]
+    if cols === nothing
+        return T(Tables.table(_to_mat([
+            _sum_row(row, LC.coefficients) for row in Tables.rows(table)
+        ])))
+    else
+        return T(Tables.table(_to_mat([
+            _sum_row([row[cname] for cname in cols], LC.coefficients)
+            for row in Tables.rows(table)
+        ])))
+    end
 end