--- a/test/linear_combination.jl
+++ b/test/linear_combination.jl
@@ -175,10 +175,12 @@
             @test lc(nt) == expected
         end
 
-        @testset "dims not supported" begin
+        @testset "custom header" begin
             nt = (a = [1, 2, 3], b = [4, 5, 6])
             lc = LinearCombination([1, -1])
-            @test_throws MethodError FeatureTransforms.apply(nt, lc; dims=1)
+            expected = (x = [-3, -3, -3],)
+            @test FeatureTransforms.apply(nt, lc; header=[:x]) == expected
+            @test lc(nt; header=[:x]) == expected
         end
 
         @testset "dimension mismatch" begin
@@ -214,10 +216,12 @@
             @test lc(df) == DataFrame(:Column1 => [-3, -3, -3])
         end
 
-        @testset "dims not supported" begin
+        @testset "custom header" begin
             df = DataFrame(:a => [1, 2, 3], :b => [4, 5, 6])
             lc = LinearCombination([1, -1])
-            @test_throws MethodError FeatureTransforms.apply(df, lc; dims=1)
+            expected = DataFrame(:x => [-3, -3, -3])
+            @test FeatureTransforms.apply(df, lc; header=[:x]) == expected
+            @test lc(df; header=[:x]) == expected
         end
 
         @testset "dimension mismatch" begin