--- a/test/power.jl
+++ b/test/power.jl
@@ -95,31 +95,29 @@
 
     @testset "NamedTuple" begin
         nt = (a = [1, 2, 3], b = [4, 5, 6])
-        expected = [[1, 8, 27], [64, 125, 216]]
-        expected_nt = (a = [1, 8, 27], b = [64, 125, 216])
 
         @testset "all cols" begin
-            @test FeatureTransforms.apply(nt, p) == expected
-            @test p(nt) == expected
+
+            expected_nt = (Column1 = [1, 8, 27], Column2 = [64, 125, 216])
+            @test FeatureTransforms.apply(nt, p) == expected_nt
+            @test p(nt) == expected_nt
 
             _nt = deepcopy(nt)
             FeatureTransforms.apply!(_nt, p)
             @test _nt isa NamedTuple{(:a, :b)}
-            @test _nt == expected_nt
+            @test _nt == (a = [1, 8, 27], b = [64, 125, 216])
         end
 
         @testset "cols = $c" for c in (:a, :b)
-            nt_mutated = NamedTuple{(Symbol("$c"), )}((expected_nt[c], ))
-            expected_nt_mutated = merge(nt, nt_mutated)
+            expected = getproperty(nt, c) .^3
 
-            @test FeatureTransforms.apply(nt, p; cols=[c]) == [expected_nt[c]]
-            @test FeatureTransforms.apply(nt, p; cols=c) == expected_nt[c]
-            @test p(nt; cols=[c]) == [expected_nt[c]]
+            @test FeatureTransforms.apply(nt, p; cols=c) == (Column1 = expected, )
+            @test p(nt; cols=c) == (Column1 = expected, )
 
             @testset "mutating" for _c in (c, [c])
                 _nt = deepcopy(nt)
                 FeatureTransforms.apply!(_nt, p; cols=_c)
-                @test _nt == expected_nt_mutated
+                @test _nt == merge(nt, NamedTuple{(c,)}((expected,)))
                 @test _nt isa NamedTuple
             end
         end
@@ -127,23 +125,23 @@
 
     @testset "DataFrame" begin
         df = DataFrame(:a => [1, 2, 3], :b => [4, 5, 6])
-        expected_df = DataFrame(:a => [1, 8, 27], :b => [64, 125, 216])
-        expected = [expected_df.a, expected_df.b]
 
         @testset "all cols" begin
-            @test FeatureTransforms.apply(df, p) == expected
-            @test p(df) == expected
+            expected_df = DataFrame(:Column1 => [1, 8, 27], :Column2 => [64, 125, 216])
+            @test FeatureTransforms.apply(df, p) == expected_df
+            @test p(df) == expected_df
 
             _df = deepcopy(df)
             FeatureTransforms.apply!(_df, p)
             @test _df isa DataFrame
-            @test _df == expected_df
+            @test _df == DataFrame(:a => [1, 8, 27], :b => [64, 125, 216])
         end
 
         @testset "cols = $c" for c in (:a, :b)
-            @test FeatureTransforms.apply(df, p; cols=[c]) == [expected_df[!, c]]
-            @test FeatureTransforms.apply(df, p; cols=c) == expected_df[!, c]
-            @test p(df; cols=[c]) == [expected_df[!, c]]
+            expected = getproperty(df, c) .^ 3
+            @test FeatureTransforms.apply(df, p; cols=[c]) == DataFrame(:Column1=>expected)
+            @test FeatureTransforms.apply(df, p; cols=c) == DataFrame(:Column1=>expected)
+            @test p(df; cols=[c]) == DataFrame(:Column1=>expected)
         end
     end
 end