Closes #55 

Alternative to #60 (closed)

This circumvents #55 altogether by returning the same type of output for a given input.
Right now this is guaranteed for  `AbstractArray` but _not_ for Tables. So to me it makes sense to actually support this rather than continuing to special-case tables and then special-casing what `cols` does (++complexity, ++inconsistency).

We were against this initially because we didn't want to come up with a naming scheme for the returned Table.
However, [`Tables.table`](https://tables.juliadata.org/stable/#Tables.table) allows for default `headers` which seem perfectly reasonable for us to extend as part of supporting Tables. We could even expose this kwarg to the user so they can pass in their own `header` names if they want.

Furthermore, we can reconstruct the original input type thanks to [`Tables.materializer`](https://tables.juliadata.org/stable/#Tables.materializer) (thanks @morris25 for bringing my attention to this!) which makes this much easier and is actually intended precisely for this purpose in applying transformations:

```
For a table input, return the "sink" function or "materializing" function that can take a Tables.jl-compatible table input 
and make an instance of the table type. This enables "transform" workflows that take table inputs, apply transformations, 
potentially converting the table to a different form, and end with producing a table of the same type as the original input. 
The default materializer is Tables.columntable, which converts any table input into a NamedTuple of Vectors.
```

Benchmarking: note that it's less performant than main probably because we are constructing Tables, but at least it's type-stable.
```julia
julia> using DataFrames, FeatureTransforms
[ Info: Precompiling FeatureTransforms [8fd68953-04b8-4117-ac19-158bf6de9782]

julia> df = DataFrame(rand(100, 100));

julia> p = Power(3);

# main
julia> @time FeatureTransforms.apply(df, p, cols=[:x1]);
  0.153022 seconds (399.12 k allocations: 21.648 MiB)
  0.000019 seconds (8 allocations: 1.203 KiB)

julia> @time FeatureTransforms.apply(df, p, cols=:x1);
  0.006716 seconds (7.00 k allocations: 427.577 KiB)
  0.000011 seconds (5 allocations: 992 bytes)

# this branch
julia> @time FeatureTransforms.apply(df, p, cols=[:x1]);
  0.398376 seconds (1.29 M allocations: 68.336 MiB, 18.14% gc time)
  0.000036 seconds (35 allocations: 5.328 KiB)

julia> @time FeatureTransforms.apply(df, p, cols=:x1);
  0.012736 seconds (8.49 k allocations: 513.706 KiB)
  0.000043 seconds (34 allocations: 5.234 KiB)

julia> @code_warntype FeatureTransforms.apply(df, p, cols=:a)
Variables
  #unused#::Core.Compiler.Const(FeatureTransforms.var"#apply##kw"(), false)
  @_2::NamedTuple{(:cols,),Tuple{Symbol}}
  @_3::Core.Compiler.Const(FeatureTransforms.apply, false)
  table::DataFrame
  t::Power
  kwargs...::Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}
  cols::Symbol
  @_8::Symbol

Body::DataFrame
1 ─       Core.NewvarNode(:(kwargs...))
│   %2  = Base.haskey(@_2, :cols)::Core.Compiler.Const(true, false)
│         %2
│         (@_8 = Base.getindex(@_2, :cols))
└──       goto #3
2 ─       Core.Compiler.Const(:(@_8 = FeatureTransforms.nothing), false)
3 ┄       (cols = @_8)
│   %8  = (:cols,)::Core.Compiler.Const((:cols,), false)
│   %9  = Core.apply_type(Core.NamedTuple, %8)::Core.Compiler.Const(NamedTuple{(:cols,),T} where T<:Tuple, false)
│   %10 = Base.structdiff(@_2, %9)::Core.Compiler.Const(NamedTuple(), false)
│         (kwargs... = Base.pairs(%10))
│   %12 = FeatureTransforms.:(var"#apply#4")(cols, kwargs..., @_3, table, t)::DataFrame
└──       return %12

```