@morris25 2021-03-22
I think this solution is preferable to #60 since we are only supporting DataFrames via the Tables interface.