@bencottier 2021-03-23
> Furthermore, we can reconstruct the original input type thanks to Tables.materializer

This is new info for me and changes my preference. In the past I've wished we had this, and the array return type stuff has just been the best we can do.

This also changes the question on the single-column syntax. Before, with #26 , it was just in the context of what kind of array to return. If everything's a Table, I don't mind there being no distinction in this syntax.

Overall I like this and it feels more like how it's meant to be than #60