@glennmoy 2021-03-23
> one caveat is that I'm still not sure what users would want here. Can we glean this from the test session?

There was nothing explicitly said about it. But I think the option to specify the column header is enough, as long as we fall back to default `Tables.table` then at least we are not asserting out opinion on it.