@glennmoy 2021-03-24
Note I haven't updated the docs yet - nor have I tested passing through a `header` kwarg except for the doctest.